+++ src/main/kotlin/com/github/iml885203/intellijgitmergeinto/EnumErrorCode.kt
@@ -1,0 +1,5 @@
+package com.github.iml885203.intellijgitmergeinto
+
+enum class EnumErrorCode {
+    UncommittedChanges
+}