+++ src/main/kotlin/com/github/iml885203/intellijgitmergeinto/MergeIntoException.kt
@@ -1,0 +1,3 @@
+package com.github.iml885203.intellijgitmergeinto
+
+class MergeIntoException(val errorCode: EnumErrorCode, override val message: String) : Throwable()